Updated all games to use floating point 0-1 when dealing with sector light levels. Part #1

+++ doomsday/plugins/doom64tc/include/p_spec.h
@@ -85,16 +85,16 @@ typedef struct {
     thinker_t       thinker;
     sector_t       *sector;
     int             count;
-    int             maxlight;
-    int             minlight;
+    float           maxlight;
+    float           minlight;
 } fireflicker_t;
 
 typedef struct {
     thinker_t       thinker;
     sector_t       *sector;
     int             count;
-    int             maxlight;
-    int             minlight;
+    float           maxlight;
+    float           minlight;
     int             maxtime;
     int             mintime;
 } lightflash_t;
@@ -103,8 +103,8 @@ typedef struct {
     thinker_t       thinker;
     sector_t       *sector;
     int             count;
-    int             maxlight;
-    int             minlight;
+    float           maxlight;
+    float           minlight;
     int             maxtime;
     int             mintime;
 } lightblink_t;
@@ -113,8 +113,8 @@ typedef struct {
     thinker_t       thinker;
     sector_t       *sector;
     int             count;
-    int             minlight;
-    int             maxlight;
+    float           minlight;
+    float           maxlight;
     int             darktime;
     int             brighttime;
 } strobe_t;
@@ -122,8 +122,8 @@ typedef struct {
 typedef struct {
     thinker_t       thinker;
     sector_t       *sector;
-    int             minlight;
-    int             maxlight;
+    float           minlight;
+    float           maxlight;
     int             direction;
 } glow_t;
 
@@ -147,7 +147,7 @@ void P_SpawnStrobeFlash(sector_t *sector, int fastOrSlow, int inSync);
 void EV_StartLightStrobing(line_t *line);
 void EV_TurnTagLightsOff(line_t *line);
 
-void EV_LightTurnOn(line_t *line, int bright);
+void EV_LightTurnOn(line_t *line, float bright);
 
 void T_Glow(glow_t *g);
 void P_SpawnGlowingLight(sector_t *sector);
